+++ example/lib/src/controller.dart
@@ -2,8 +2,10 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'package:example/src/another_app/controller.dart';
+export '/src/another_app/controller.dart';
 
-export 'package:example/src/controller/app/app_controller.dart';
+export '/src/controller/app/app_controller.dart';
 
-export 'package:example/src/controller/home/_controller.dart';
+export '/src/controller/app/common/events_controller.dart';
+
+export '/src/controller/home/_controller.dart';