+++ example/lib/src/controller/app/app_controller.dart
@@ -2,10 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:example/src/view.dart';
+import '/src/controller.dart';
+
+import '/src/view.dart';
 
 /// The 'App Level' Controller
-class ExampleAppController extends StateXController {
+class ExampleAppController extends StateXController with EventsControllerMixin {
   /// Singleton design pattern is best for Controllers.
   factory ExampleAppController() => _this ??= ExampleAppController._();
   ExampleAppController._();
@@ -37,67 +39,4 @@ class ExampleAppController extends StateXController {
 
   /// Allow for a Splash screen or not
   bool splashScreen = false;
-
-  /// The framework will call this method exactly once.
-  /// Only when the [StateX] object is first created.
-  @override
-  void initState() {
-    super.initState();
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: initState in $state');
-    }
-  }
-
-  /// The framework calls this method whenever this [StateX] object is closed.
-  @override
-  void deactivate() {
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: deactivate in $state');
-    }
-  }
-
-  /// Called when this object was [deactivate].
-  /// Rarely actually called but good to have since was deactivated.
-  /// You may have to undo what was done in [deactivate].
-  /// It was removed from the widget tree for some reason but re-inserted again.
-  @override
-  void activate() {
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: activate in $state');
-    }
-  }
-
-  /// The framework calls this method when this [StateX] object will be garbage collected
-  /// Note: YOU HAVE NO IDEA WHEN THIS WILL RUN in the Framework.
-  /// USE [deactivate] for time critical operations.[dispose] for light cleanup.
-  @override
-  void dispose() {
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: dispose in $state');
-    }
-    super.dispose();
-  }
-
-  /// The application is not currently visible to the user, not responding to
-  /// user input, and running in the background.
-  @override
-  void pausedLifecycleState() {
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: pausedLifecycleState in $state');
-    }
-  }
-
-  /// The application is visible and responding to user input.
-  @override
-  void resumedLifecycleState() {
-    if (inDebugMode) {
-      //ignore: avoid_print
-      print('############ Event: resumedLifecycleState in $state');
-    }
-  }
 }