+++ example/lib/src/controller/home/yet_another_controller.dart
@@ -2,11 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:example/src/view.dart';
+import '/src/controller.dart';
+
+import '/src/view.dart';
 
 /// Yet another Controller for demonstration purposes.
 /// Includes the mixin, StateXonErrorMixin, to supply an error handler
-class YetAnotherController extends StateXController with StateXonErrorMixin {
+class YetAnotherController extends StateXController
+    with EventsControllerMixin, StateXonErrorMixin {
   /// It's a good practice to make Controllers using the Singleton pattern
   factory YetAnotherController() => _this ??= YetAnotherController._();
   YetAnotherController._() : super();
@@ -17,7 +20,7 @@ class YetAnotherController extends StateXController with StateXonErrorMixin {
   void onError(FlutterErrorDetails details) {
     if (inDebugMode) {
       //ignore: avoid_print
-      print('############ Event: onError in YetAnotherController for $state');
+      print('============ Event: onError in YetAnotherController for $state');
     }
   }
 }